refactor: extract repo map and symbol extraction into server/work/repo-map.ts
Move generateRepoMap, filePathPriority, extractRelevantSymbols,
tokenizeDescription and related constants (REPO_MAP_MAX_LINES,
PRIORITY_DIRS, STOP_WORDS) out of WorkTaskService into standalone
pure functions that accept AstParserService as a parameter.

service.ts: 1008 → 776 lines, repo-map.ts: 243 lines.
All 5471 tests pass, tsc clean.
